applications/violet_lambda_function/requirements.txt
guessit==3.8.0
python-slugify==8.0.4
themoviedb[sync]==1.0.2
aws-lambda-powertools==3.25.0
